refactor: break up panel_space.rs into multiple smaller files & wip toplevel handling
