+++ cosmic-panel-bin/src/space/mod.rs
@@ -1,7 +1,10 @@
 //! PanelSpace is a container for all running panels, spawning each as a separate process and compositing them in a layer shell surface as configured
 //! PanelSpace *partially* implements the WrapperSpace abstraction
 
+mod layout;
 mod panel_space;
+mod popup;
+mod render;
 mod toplevel;
 mod wrapper_space;
